skills/smart-fetch/scripts/requirements.txt
curl_cffi
DrissionPage
readability-lxml
markdownify
lxml
